requirements.txt
asarPy==1.0.1
blinker==1.9.0
click==8.2.0
Cython==3.0.11
dnspython==2.7.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.8.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==24.2
pycairo==1.27.0
Pygments==2.19.1
PyGObject==3.50.0
pymongo==4.13.2
python-dotenv==1.1.1
setuptools==75.8.0.post20250203
six==1.17.0
Werkzeug==3.1.3
wheel==0.45.1
